refactor: move `LegalDisclaimer`
