+++ packages/frontend/src/pages/App/App.tsx
@@ -21,7 +21,7 @@ import { Backdrop } from '../../components/Miscellaneous/Backdrop/Backdrop'
 import { SearchBar } from '../../components/Miscellaneous/SearchBar/SearchBar'
 import { StatsPopUp } from '../../components/Miscellaneous/StatsPopUp/StatsPopUp'
 import { InfoModal } from '../../components/Modals/InfoModal/InfoModal'
-import { LegalDisclaimer } from '../../components/Modals/LegalDisclaimer/LegalDisclaimer'
+import { LegalDisclaimer } from '../../components/Modals/LegalDisclaimer'
 import { UtilModal } from '../../components/Modals/UtilModal/UtilModal'
 import { ViewedReportsProvider } from '../../contexts/ViewedReportsContext'
 import { sendAnalyticsEvent, sendSavedEvents } from '../../hooks/useAnalytics'
